Refactor imports for @hedvig-ui/OrbIndicator
