+++ shared/hedvig-ui/index.ts
@@ -42,6 +42,8 @@ import { CustomInputProps, Input } from './Input/input'
 import { JsonSchemaForm } from './JsonSchemaForm/json-schema-form'
 import { List, ListItem } from './List/list'
 import { Loadable } from './Loadable/loadable'
+import { FlagOrbIndicator, OrbIndicator } from './OrbIndicator/orb-indicator'
+import { Popover } from './Popover/popover'
 
 export {
   Badge,
@@ -81,4 +83,7 @@ export {
   ListItem,
   JsonSchemaForm,
   Loadable,
+  Popover,
+  OrbIndicator,
+  FlagOrbIndicator,
 }
