+++ src/components/member/index.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled'
+import { Popover } from '@hedvig-ui'
 import { Member } from 'api/generated/graphql'
 import { memberPagePanes } from 'components/member/tabs'
 import { ChatPane } from 'components/member/tabs/ChatPane'
 import copy from 'copy-to-clipboard'
-import { Popover } from 'hedvig-ui/popover'
 import React, { useContext, useEffect, useState } from 'react'
 import { toast } from 'react-hot-toast'
 import { RouteComponentProps, useHistory } from 'react-router'
